Add constructor "Unshared" to type SharedTerm s.
This allows us to defer the hash consing of shared terms if we want.
It also means that we can build unshared terms in pure code, without
using a SharedContext.

Eventually this should allow us to get rid of the old unshared Term
type, because the Unshared constructor should let us do everything
that we previously needed Term for.