Use 'Control.Monad.Random' instead of manual generator-state threading.
The implementations become much simpler but the type signatures get
worse. Could shorten the signatures by introducing a constraint
synonym for the common

  (Applicative m, Functor m, MonadRandom m)

constraint.

Added, and used in 'Builtins', an interface function 'scRunTestsTFIO'
that hides the random state stuff completely from the user.

Uncovered an embarrassing bug I introduced in the original
translation: I had translated

  2^w-1

to

  2^(unNat w - 1)

instead of

  2^(unNat w) - 1

:P