Better error message for SBV 'codegen' command if the result is a list of bits
The C code generation in the SBV library does not support concatenation
of a list of bits to form a result value, so if the primitives used in
a function produce a result value that is represented as a list of bits,
then the code generation must fail. We should be able to extend the set
of translatable programs by reimplementing some primitives in the SBV
backend to preserve the word-level representation as much as possible.

This partially addresses GaloisInc/saw-script#7.