Add the `expByNat` primitive to SAWCore.  This is used to exponentiate and arbitrary type by a natural number value.
This is made a separate primitive instead of relying on the recursor
for `Nat` for two reasons.  First, SAWCore is quite fragile in some
ways when `Nat` values are created via `bvToNat`.  The Nat recursor
will panic if it encounters such a value.  Second, we want to implement
the exp primitive via square-and-multiply rather than directly
iterating multiplication.