saw-core-coq: clean _CoqProject file
The _CoqProject file used to contain a lot of extraneous files related to my
work on S2N, my work with Talia Ringer, and an attempt at work on BESSPIN.
These are obsolete and irrelevant to saw-core-coq and should be located
elsewhere.

This commit also re-enables the compilation of some saw-core-coq files, notably
CryptolPrimitivesForSAWCore which is necessary for extraction from SAW, and
CoqVectorsExtra which may not be very useful yet.

This also remove all unnecessary Coq and Cryptol files.